+++ src/vlite/components/big-play/assets/scripts/big-play.ts
@@ -4,6 +4,6 @@ import svgBigPlay from 'shared/assets/svgs/big-play.svg'
  * Big play template
  * @returns {String} Generated HTML
  */
-export default function (): string {
+export default function bigPlay(): string {
 	return `<button class="v-bigPlay v-controlButton">${svgBigPlay}</button>`
 }