+++ src/vlite/components/overlay/assets/scripts/overlay.ts
@@ -2,6 +2,6 @@
  * Overlay template
  * @returns {String} Generated HTML
  */
-export default function (): string {
+export default function overlay(): string {
 	return '<div class="v-overlay"></div>'
 }