+++ src/vlite/components/poster/assets/scripts/poster.ts
@@ -4,7 +4,7 @@
  * @param {Object} options.posterUrl Poster url
  * @returns {String} Generated HTML
  */
-export default function ({ posterUrl = '' }: { posterUrl: string }): string {
+export default function poster({ posterUrl = '' }: { posterUrl: string }): string {
 	const posterStyle = posterUrl && ` style="background-image: url(${posterUrl})"`
 	return `<div class="v-poster v-active"${posterStyle}></div>`
 }